changes schedule to use time, creates schedule view

+++ db/migrate/20180618215117_create_schedules.rb
@@ -1,13 +1,12 @@
 class CreateSchedules < ActiveRecord::Migration[5.2]
   def change
     create_table :schedules do |t|
-      t.integer :time
+      t.float :time
       t.integer :user_id
       t.integer :habit_id
       t.string :frequency
       t.datetime :start_date
       t.datetime :end_date
-
       t.timestamps
     end
   end
